fix bug of pixels of treeCut was integer

+++ app/model/Query.java
@@ -12,6 +12,6 @@ public class Query {
     public String indexType; // "KDTree"(default) / "GridIndex"
     public boolean treeCut;  // true / false
     public String measure; // for treeCut: "avg" / "min" / "max" distance of children
-    public int pixels; // for treeCut: number of pixels for differentiable distance in tree-cut algorithm
+    public double pixels; // for treeCut: number of pixels for differentiable distance in tree-cut algorithm
     public boolean bipartite; // for treeCut: true / false
 }
